Update harshad number C implementation: Add tests and more descriptions
