+++ 10_Harshad_Number/C/jb1717/10_Harshad_Number.h
@@ -1,0 +1,6 @@
+#ifndef HARSHAD_NUMBER
+# define HARSHAD_NUMBER
+
+int isHarshad(int n);
+
+#endif
